✨ Rename GithubStars component file for consistency and update import paths
